- create an examples directory to store pre-packaged reference forms - move KT form v8 to the examples directory (out from db_setup.py)
